Reorganized files to match contributing guidelines
