+++ src/components/Section/ItemBlock.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 
-import { SectionItem } from './Section';
+import { SectionItem } from '../Section';
 
-import './styles/ItemBlock.scss';
+import '../styles/Section.scss';
 
 interface ItemBlockProps {
   data: SectionItem;
